add working Keyman component
