--- packages/parts/src/lib/entries/keyboards/Keyman.svelte
@@ -1,8 +1,0 @@
-<svelte:head>
-  <script src="https://s.keyman.com/kmw/engine/15.0.18/keymanweb.js"></script>
-  <script src="https://s.keyman.com/kmw/engine/15.0.18/kmwuibutton.js"></script>
-  <!-- float, button, toggle, toolbar -->
-  <!-- https://s.keyman.com/kmw/engine/16.0.59/keymanweb.js -->
-</svelte:head>
-
-Hello
