+++ packages/parts/src/lib/entries/keyboards/ipa/vowel_trapezoid.gif
