--- packages/parts/src/lib/entries/keyboards/keyman.css
@@ -1,10 +1,0 @@
-.kmw-key:hover {
-  background: #dcdcdc;
-}
-
-.kmw-title-bar,
-.desktop .kmw-osk-inner-frame,
-.kmw-footer {
-  background-color: rgb(68, 68, 68) !important;
-  border-color: rgb(68, 68, 68);
-}
